Only overwrite version.c if content changed
If we don't check for this, make will always believe a rebuild is needed
because of this dynamically generated file, which ruins the idea of
dependency tree.

To do so a temporary file is generated and its content is diffed against
the existing version.c. If the two are identical, no file is overwritten.