Add a command to turn on or off the bus power
This commit adds two commands to enable or disable the CAN bus power
from the host computer. As no commands exists in the "official" SLCAN
protocol, I used a proprietary extension that does not clash with the
existing ones.

This type of command can be useful in scenarios where human intervention
is not possible / desired, such as CI systems.